Refactored.

+++ Game/src/UI/FTextField.h
@@ -8,19 +8,19 @@ namespace wce
 
 	// Constructors and Destructor:
 
-		 FTextField ();
-		~FTextField ();
+		 FTextField    ();
+		~FTextField    ();
 
 	// Functions:
 
-		void Draw (class FScreenBuffer& ScreenBuffer);
+		void Draw    (class FScreenBuffer& ScreenBuffer);
 
 	// Accessors:
 
 		const COORD&        GetPosition     () const;
 		const WORD&         GetAttribute    () const;
 		const std::wstring& GetText         () const;
-		uint64              GetLength       () const;
+		size_t              GetLength       () const;
 
 	// Modifiers:
